requirements.txt
fastapi==0.111.0
langchain==0.2.6
langchain_community==0.2.6
langchain_core==0.2.10
langchain_milvus==0.1.1
langchain_mistralai==0.1.8
peewee==3.17.5
pydantic==2.7.4
pytest==8.2.2
PyYAML==6.0.1
toml==0.10.2
pypdf==4.2.0
docx2txt==0.8
GitPython==3.1.43
uvicorn==0.30.1
psycopg2-binary==2.9.9
coverage==7.5.3
chromadb==0.5.3